requirements.txt
Flask
requests
python-dotenv
langchain
langchain-openai
langchain-core
rich